+++ examples/boids.rs
@@ -129,7 +129,7 @@ fn main() -> RendererResult<()> {
             }
         }
 
-        if let Some(window) = renderer.get_window(window_id) {
+        if let Some(mut window) = renderer.get_window(window_id) {
             window.background(Color::BLACK);
 
             // TODO: should probably optimize this loop so we can handle more boids