+++ examples/drawing.rs
@@ -17,7 +17,7 @@ fn main() -> RendererResult<()> {
             }
         }
 
-        if let Some(window) = renderer.get_window(window_id) {
+        if let Some(mut window) = renderer.get_window(window_id) {
             window.background(Color::GRAY);
 
             // imperative, state-machine API
@@ -28,10 +28,10 @@ fn main() -> RendererResult<()> {
             Rect::at(250., 500.)
                 .size(250., 250.)
                 .fill(Color::BLUE)
-                .draw(window);
+                .draw(&mut window);
 
             mouse_ellipse.position(window.get_mouse_x(), window.get_mouse_y());
-            mouse_ellipse.draw(window);
+            mouse_ellipse.draw(&mut window);
         }
 
         renderer.flush()?;