+++ examples/font_fallback.rs
@@ -20,7 +20,7 @@ fn main() -> RendererResult<()> {
             }
         }
 
-        if let Some(window) = renderer.get_window(window_id) {
+        if let Some(mut window) = renderer.get_window(window_id) {
             window.background(Color::BLACK);
 
             window.font_size(32.);