+++ examples/rich_text.rs
@@ -40,7 +40,7 @@ fn main() -> RendererResult<()> {
             }
         }
 
-        if let Some(window) = renderer.get_window(window_id) {
+        if let Some(mut window) = renderer.get_window(window_id) {
             window.background(Color::BLACK);
 
             // set the text alignment (where the origin is)